+++ src/Sleet/DestroyAppCommand.cs
@@ -22,19 +22,12 @@ namespace Sleet
                 CommandOptionType.SingleValue);
 
             var verbose = cmd.Option(Constants.VerboseOption, Constants.VerboseDesc, CommandOptionType.NoValue);
+            var propertyOptions = cmd.Option(Constants.PropertyOption, Constants.PropertyDescription, CommandOptionType.MultipleValue);
 
             cmd.HelpOption(Constants.HelpOption);
 
-            var required = new List<CommandOption>()
-            {
-                sourceName
-            };
-
             cmd.OnExecute(async () =>
             {
-                // Validate parameters
-                CmdUtils.VerifyRequiredOptions(required.ToArray());
-
                 // Init logger
                 Util.SetVerbosity(log, verbose.HasValue());
 
@@ -42,7 +35,7 @@ namespace Sleet
                 using (var cache = new LocalCache())
                 {
                     // Load settings and file system.
-                    var settings = LocalSettings.Load(optionConfigFile.Value());
+                    var settings = LocalSettings.Load(optionConfigFile.Value(), SettingsUtility.GetPropertyMappings(propertyOptions.Values));
                     var fileSystem = Util.CreateFileSystemOrThrow(settings, sourceName.Value(), cache);
 
                     // Delete files